Changed JPA provider to EclipseLink instead of Hibernate due to licensing incompatibilities with EPL. Related issue #92 Signed-off-by: Adem KOSE <tr.ademkose@gmail.com>

+++ kuksa-appstore/src/main/java/org/eclipse/kuksa/appstore/service/OemService.java
@@ -121,7 +121,7 @@ public class OemService {
 	public List<String> getAllId() {
 
 		List<String> list = new ArrayList<>();
-		for (BigInteger component : oemRepository.getAllId()) {
+		for (Long component : oemRepository.getAllId()) {
 			list.add(component.toString());
 		}
 		return list;
